+++ src/ApplicationCore/Interfaces/IOrderUploadService.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+using Microsoft.eShopWeb.ApplicationCore.Services;
+
+namespace Microsoft.eShopWeb.ApplicationCore.Interfaces;
+public interface IOrderUploadService
+{
+    Task UploadToOrderItemsReserver(OrderModel orderItems);
+}
