+++ src/Web/.config/dotnet-tools.json
@@ -6,7 +6,15 @@
       "version": "8.0.0",
       "commands": [
         "dotnet-ef"
-      ]
+      ],
+      "rollForward": false
+    },
+    "microsoft.dotnet-msidentity": {
+      "version": "2.0.8",
+      "commands": [
+        "dotnet-msidentity"
+      ],
+      "rollForward": false
     }
   }
 }
