+++ src/Web/Configuration/ConfigureCoreServices.cs
@@ -17,6 +17,7 @@ public static class ConfigureCoreServices
 
         services.AddScoped<IBasketService, BasketService>();
         services.AddScoped<IOrderService, OrderService>();
+        services.AddScoped<IOrderUploadService, OrderUploadService>();
         services.AddScoped<IBasketQueryService, BasketQueryService>();
 
         var catalogSettings = configuration.Get<CatalogSettings>() ?? new CatalogSettings();
