+++ src/Web/Pages/Basket/Checkout.cshtml.cs
@@ -1,4 +1,5 @@
-﻿using System.Text;
+﻿using System.Runtime.CompilerServices;
+using System.Text;
 using System.Text.Json;
 using Ardalis.GuardClauses;
 using Azure.Core;
@@ -9,6 +10,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate;
 using Microsoft.eShopWeb.ApplicationCore.Exceptions;
 using Microsoft.eShopWeb.ApplicationCore.Interfaces;
+using Microsoft.eShopWeb.ApplicationCore.Services;
 using Microsoft.eShopWeb.Infrastructure.Identity;
 using Microsoft.eShopWeb.Web.Interfaces;
 
@@ -22,18 +24,21 @@ public class CheckoutModel : PageModel
     private readonly IOrderService _orderService;
     private string? _username = null;
     private readonly IBasketViewModelService _basketViewModelService;
+    private readonly IOrderUploadService _orderUploadService;
     private readonly IAppLogger<CheckoutModel> _logger;
 
     public CheckoutModel(IBasketService basketService,
         IBasketViewModelService basketViewModelService,
         SignInManager<ApplicationUser> signInManager,
         IOrderService orderService,
+        IOrderUploadService orderUploadService,
         IAppLogger<CheckoutModel> logger)
     {
         _basketService = basketService;
         _signInManager = signInManager;
         _orderService = orderService;
         _basketViewModelService = basketViewModelService;
+        _orderUploadService = orderUploadService;
         _logger = logger;
     }
 
@@ -64,7 +69,7 @@ public class CheckoutModel : PageModel
                 Items = items.Select(x => new Item { Id = x.Id, Quantity = x.Quantity })
             };
 
-            await UploadToOrderItemsReserver(orderItems);
+            await _orderUploadService.UploadToOrderItemsReserver(orderItems);
 
             await _basketService.SetQuantities(BasketModel.Id, updateModel);
             await _orderService.CreateOrderAsync(BasketModel.Id, new Address("123 Main St.", "Kent", "OH", "United States", "44240"));
@@ -80,18 +85,18 @@ public class CheckoutModel : PageModel
         return RedirectToPage("Success");
     }
 
-    private async Task UploadToOrderItemsReserver(OrderModel orderItems)
-    {
-        using (HttpClient httpClient = new HttpClient()) { 
+    //private async Task UploadToOrderItemsReserver(OrderModel orderItems)
+    //{
+    //    using (HttpClient httpClient = new HttpClient()) { 
 
-            var functionurl = "https://eshoponwebtaskfunctionapp.azurewebsites.net/api/OrderModelUpload?code=-vr_jqDEcT_lMruys_-yaUPvrxV1H78BYyQJQk4y8qWgAzFuom9LlA%3D%3D";
+    //        var functionurl = "https://eshoponwebtaskfunctionapp.azurewebsites.net/api/OrderModelUpload?code=-vr_jqDEcT_lMruys_-yaUPvrxV1H78BYyQJQk4y8qWgAzFuom9LlA%3D%3D";
 
-            //var content = new StringContent(JsonSerializer.Serialize(orderItems), Encoding.UTF8, "application/json");
+    //        //var content = new StringContent(JsonSerializer.Serialize(orderItems), Encoding.UTF8, "application/json");
 
-            HttpResponseMessage response = httpClient.PostAsJsonAsync(functionurl, orderItems).Result;
-            _logger.LogInformation(response.Content.ToString());
-        }
-    }
+    //        HttpResponseMessage response = httpClient.PostAsJsonAsync(functionurl, orderItems).Result;
+    //        _logger.LogInformation(response.Content.ToString());
+    //    }
+    //}
 
     private async Task SetBasketModelAsync()
     {
@@ -120,17 +125,4 @@ public class CheckoutModel : PageModel
         cookieOptions.Expires = DateTime.Today.AddYears(10);
         Response.Cookies.Append(Constants.BASKET_COOKIENAME, _username, cookieOptions);
     }
-}
-
-public class OrderModel
-{
-    public int OrderId { get; set; }
-    public string UserId { get; set; }
-    public IEnumerable<Item> Items { get; set; }
-}
-
-public class Item
-{
-    public int Id { get; set; }
-    public int Quantity { get; set; }
 }
