+++ src/Web/Properties/serviceDependencies.eshoponwebfuntask - Web Deploy.json
@@ -1,0 +1,31 @@
+{
+  "dependencies": {
+    "identityapp1": {
+      "connectionInfo": "App registration: EShopOnWebTask, Tenant: EPAM (EPAM.onmicrosoft.com - Microsoft Entra ID)",
+      "type": "identityapp.default"
+    },
+    "mssql1": {
+      "serviceConnectorResourceId": "/subscriptions/[parameters('subscriptionId')]/resourcegroups/[parameters('resourceGroupName')]/providers/Microsoft.Web/sites/eshoponwebfuntask/providers/Microsoft.ServiceLinker/linkers/ConnectionStringsCatalogConnection_09BBBA4DEB",
+      "secretStore": "AzureAppSettings",
+      "resourceId": "/subscriptions/[parameters('subscriptionId')]/resourceGroups/[parameters('resourceGroupName')]/providers/Microsoft.Sql/servers/eshoponwebservernew/databases/eshoponwebCatalog_db",
+      "type": "mssql.azure",
+      "connectionId": "ConnectionStrings:CatalogConnection",
+      "dynamicId": null
+    },
+    "mssql2": {
+      "serviceConnectorResourceId": "/subscriptions/[parameters('subscriptionId')]/resourcegroups/[parameters('resourceGroupName')]/providers/Microsoft.Web/sites/eshoponwebfuntask/providers/Microsoft.ServiceLinker/linkers/ConnectionStringsIdentityConnection_7AC3A68AD4",
+      "secretStore": "AzureAppSettings",
+      "resourceId": "/subscriptions/[parameters('subscriptionId')]/resourceGroups/[parameters('resourceGroupName')]/providers/Microsoft.Sql/servers/eshoponwebservernew/databases/eshoponwebIdentity_db",
+      "type": "mssql.azure",
+      "connectionId": "ConnectionStrings:IdentityConnection",
+      "dynamicId": null
+    },
+    "secrets1": {
+      "serviceConnectorResourceId": "/subscriptions/[parameters('subscriptionId')]/resourcegroups/[parameters('resourceGroupName')]/providers/Microsoft.Web/sites/eshoponwebfuntask/providers/Microsoft.ServiceLinker/linkers/AZURE_KEY_VAULT_ENDPOINT_2FA4A065A1",
+      "secretStore": "AzureAppSettings",
+      "resourceId": "/subscriptions/[parameters('subscriptionId')]/resourceGroups/[parameters('resourceGroupName')]/providers/Microsoft.KeyVault/vaults/eshoponwebWebvault",
+      "type": "secrets.keyVault",
+      "connectionId": "AZURE_KEY_VAULT_ENDPOINT"
+    }
+  }
+}
