+++ src/Web/Shared/LoginDisplay.razor
@@ -1,0 +1,9 @@
+<AuthorizeView>
+    <Authorized>
+        @* Hello, @context.User.Identity?.Name! *@
+        <a href="MicrosoftIdentity/Account/SignOut">Log out</a>
+    </Authorized>
+    <NotAuthorized>
+        <a href="MicrosoftIdentity/Account/SignIn">Log in</a>
+    </NotAuthorized>
+</AuthorizeView>
