+++ src/Web/appsettings.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "baseUrls": {
     "apiBase": "https://localhost:5099/api/",
     "webBase": "https://localhost:44315/"
@@ -16,5 +16,13 @@
       "System": "Warning"
     },
     "AllowedHosts": "*"
+  },
+  "AzureAd": {
+    "Instance": "https://login.microsoftonline.com/",
+    "Domain": "EPAM.onmicrosoft.com",
+    "TenantId": "b41b72d0-4e9f-4c26-8a69-f949f367c91d",
+    "ClientId": "b827c9f8-178b-482d-ad00-737724b2203f",
+    "CallbackPath": "/signin-oidc",
+    "Scopes": ""
   }
 }
